adds sync

+++ backend/modules/sync/utils/googleutils.py
@@ -16,7 +16,10 @@ from modules.sync.dto.inputs import (
 )
 from modules.sync.repository.sync_files import SyncFiles
 from modules.sync.service.sync_service import SyncService, SyncUserService
-from modules.sync.utils.list_files import get_google_drive_files
+from modules.sync.utils.list_files import (
+    get_google_drive_files,
+    get_google_drive_files_by_id,
+)
 from modules.sync.utils.upload import upload_file
 from modules.upload.service.upload_file import check_file_exists
 from pydantic import BaseModel, ConfigDict
@@ -57,10 +60,11 @@ class GoogleSyncUtils(BaseModel):
             logger.info("Google Drive credentials refreshed")
             # Updating the credentials in the database
 
-        try:
-            service = build("drive", "v3", credentials=creds)
-            downloaded_files = []
-            for file in files:
+        service = build("drive", "v3", credentials=creds)
+        downloaded_files = []
+        for file in files:
+            logger.info("🔥🔥🔥🔥: %s", file)
+            try:
                 file_id = file["id"]
                 file_name = file["name"]
                 mime_type = file["mime_type"]
@@ -68,7 +72,8 @@ class GoogleSyncUtils(BaseModel):
                 # Convert Google Docs files to appropriate formats before downloading
                 if mime_type == "application/vnd.google-apps.document":
                     logger.debug(
-                        "Converting Google Docs file with file_id: %s to DOCX.", file_id
+                        "Converting Google Docs file with file_id: %s to DOCX.",
+                        file_id,
                     )
                     request = service.files().export_media(
                         fileId=file_id,
@@ -119,7 +124,7 @@ class GoogleSyncUtils(BaseModel):
 
                 # Check if the file already exists in the storage
                 if check_file_exists(brain_id, file_name):
-                    logger.info("🔥 File already exists in the storage: %s", file_name)
+                    logger.debug("🔥 File already exists in the storage: %s", file_name)
 
                     self.storage.remove_file(brain_id + "/" + file_name)
                     BrainsVectors().delete_file_from_brain(brain_id, file_name)
@@ -129,13 +134,15 @@ class GoogleSyncUtils(BaseModel):
                     filename=file_name,
                 )
 
-                await upload_file(to_upload_file, brain_id, current_user)
-
                 # Check if the file already exists in the database
                 existing_files = self.sync_files_repo.get_sync_files(sync_active_id)
                 existing_file = next(
                     (f for f in existing_files if f.path == file_name), None
                 )
+                supported = False
+                if (existing_file and existing_file.supported) or not existing_file:
+                    supported = True
+                    await upload_file(to_upload_file, brain_id, current_user)  # type: ignore
 
                 if existing_file:
                     # Update the existing file record
@@ -143,6 +150,7 @@ class GoogleSyncUtils(BaseModel):
                         existing_file.id,
                         SyncFileUpdateInput(
                             last_modified=modified_time,
+                            supported=supported,
                         ),
                     )
                 else:
@@ -153,16 +161,41 @@ class GoogleSyncUtils(BaseModel):
                             syncs_active_id=sync_active_id,
                             last_modified=modified_time,
                             brain_id=brain_id,
+                            supported=supported,
                         )
                     )
 
-                downloaded_files.append(file_name)
-            return {"downloaded_files": downloaded_files}
-        except HttpError as error:
-            logger.error(
-                "An error occurred while downloading Google Drive files: %s", error
-            )
-            return {"error": f"An error occurred: {error}"}
+                    downloaded_files.append(file_name)
+            except HttpError as error:
+                logger.error(
+                    "An error occurred while downloading Google Drive files: %s",
+                    error,
+                )
+                # Check if the file already exists in the database
+                existing_files = self.sync_files_repo.get_sync_files(sync_active_id)
+                existing_file = next(
+                    (f for f in existing_files if f.path == file["name"]), None
+                )
+                # Update the existing file record
+                if existing_file:
+                    self.sync_files_repo.update_sync_file(
+                        existing_file.id,
+                        SyncFileUpdateInput(
+                            supported=False,
+                        ),
+                    )
+                else:
+                    # Create a new file record
+                    self.sync_files_repo.create_sync_file(
+                        SyncFileInput(
+                            path=file["name"],
+                            syncs_active_id=sync_active_id,
+                            last_modified=file["last_modified"],
+                            brain_id=brain_id,
+                            supported=False,
+                        )
+                    )
+        return {"downloaded_files": downloaded_files}
 
     async def sync(self, sync_active_id: int, user_id: str):
         """
@@ -183,8 +216,9 @@ class GoogleSyncUtils(BaseModel):
 
         # Check if the sync is due
         last_synced = sync_active.get("last_synced")
+        force_sync = sync_active.get("force_sync", False)
         sync_interval_minutes = sync_active.get("sync_interval_minutes", 0)
-        if last_synced:
+        if last_synced and not force_sync:
             last_synced_time = datetime.fromisoformat(last_synced).astimezone(
                 timezone.utc
             )
@@ -229,12 +263,26 @@ class GoogleSyncUtils(BaseModel):
             sync_active_id,
         )
 
-        # Get the folder id from the settings from sync_active
         settings = sync_active.get("settings", {})
         folders = settings.get("folders", [])
-        files = get_google_drive_files(
-            sync_user["credentials"], folder_id=folders[0] if folders else None
-        )
+        files_to_download = settings.get("files", [])
+        files = []
+        files_metadata = []
+        if len(folders) > 0:
+            files = []
+            for folder in folders:
+                files.extend(
+                    get_google_drive_files(
+                        sync_user["credentials"],
+                        folder_id=folder,
+                        recursive=True,
+                    )
+                )
+        if len(files_to_download) > 0:
+            files_metadata = get_google_drive_files_by_id(
+                sync_user["credentials"], files_to_download
+            )
+        files = files + files_metadata  # type: ignore
         if "error" in files:
             logger.error(
                 "Failed to download files from Google Drive for sync_active_id: %s",
@@ -244,15 +292,21 @@ class GoogleSyncUtils(BaseModel):
 
         # Filter files that have been modified since the last sync
         last_synced_time = datetime.fromisoformat(last_synced) if last_synced else None
+
         files_to_download = [
             file
-            for file in files.get("files", [])
+            for file in files
             if not file["is_folder"]
             and (
-                not last_synced_time
-                or datetime.fromisoformat(file["last_modified"]) > last_synced_time
+                (
+                    not last_synced_time
+                    or datetime.fromisoformat(file["last_modified"]) > last_synced_time
+                )
+                or not check_file_exists(sync_active["brain_id"], file["name"])
             )
         ]
+
+        logger.error(files_to_download)
 
         downloaded_files = await self._upload_files(
             sync_user["credentials"],
@@ -261,17 +315,14 @@ class GoogleSyncUtils(BaseModel):
             sync_active["brain_id"],
             sync_active_id,
         )
-        if "error" in downloaded_files:
-            logger.error(
-                "Failed to download files from Google Drive for sync_active_id: %s",
-                sync_active_id,
-            )
-            return None
 
         # Update the last_synced timestamp
         self.sync_active_service.update_sync_active(
             sync_active_id,
-            SyncsActiveUpdateInput(last_synced=datetime.now().astimezone().isoformat()),
+            SyncsActiveUpdateInput(
+                last_synced=datetime.now().astimezone().isoformat(),
+                force_sync=False,
+            ),
         )
         logger.info(
             "Google Drive sync completed for sync_active_id: %s", sync_active_id
